change to long for get byte of index

+++ src/ngram_btree_index.hpp
@@ -26,7 +26,7 @@ class NGramBtreeIndex : public NGramIndex {
 
     size_t get_num_keys() const override { return k_index_keys_.size(); }
 
-    size_t get_bytes_used() const override;
+    long get_bytes_used() const override;
 
     bool get_all_idxs(const std::string & reg, std::vector<size_t> & container) const {
         std::cerr << "should not be called" << std::endl;
